requirements.txt
asgiref==3.4.1
awsebcli==3.20.0
backports.entry-points-selectable==1.1.0
botocore==1.20.112
cement==2.8.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.3
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
future==0.16.0
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
pathspec==0.5.9
platformdirs==2.2.0
psycopg2==2.9.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.25.11
virtualenv==20.7.0
wcwidth==0.1.9
whitenoise==5.3.0
